+++ lib/screens/settings.dart
@@ -1,15 +1,13 @@
-import 'package:adaptive_dialog/adaptive_dialog.dart';
 import 'package:every_door/constants.dart';
 import 'package:every_door/providers/changes.dart';
 import 'package:every_door/providers/changeset_tags.dart';
 import 'package:every_door/providers/editor_settings.dart';
 import 'package:every_door/providers/geolocation.dart';
-import 'package:every_door/providers/need_update.dart';
 import 'package:every_door/providers/notes.dart';
 import 'package:every_door/providers/osm_auth.dart';
-import 'package:every_door/providers/osm_data.dart';
 import 'package:every_door/screens/settings/about.dart';
 import 'package:every_door/screens/settings/account.dart';
+import 'package:every_door/screens/settings/caches.dart';
 import 'package:every_door/screens/settings/changes.dart';
 import 'package:every_door/screens/settings/changeset_pane.dart';
 import 'package:every_door/screens/settings/imagery.dart';
@@ -17,19 +15,15 @@ import 'package:every_door/screens/settings/language.dart';
 import 'package:every_door/screens/settings/plugins.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_dropdown_alert/alert_controller.dart';
-import 'package:flutter_dropdown_alert/model/data_alert.dart';
 import 'package:settings_ui/settings_ui.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:every_door/generated/l10n/app_localizations.dart' show AppLocalizations;
-import 'package:intl/intl.dart';
 
 class SettingsPage extends ConsumerWidget {
   const SettingsPage();
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final osmData = ref.watch(osmDataProvider);
     final login = ref.watch(authProvider)?.displayName;
     final editorSettings = ref.watch(editorSettingsProvider);
     final forceLocation = ref.watch(forceLocationProvider);
@@ -38,17 +32,6 @@ class SettingsPage extends ConsumerWidget {
 
     final haveChanges = ref.watch(changesProvider).length > 0;
     final haveNotes = ref.watch(notesProvider).haveChanges;
-
-    final purgeAll = osmData.obsoleteLength == 0;
-    NumberFormat numFormat;
-    try {
-      numFormat = NumberFormat.compact(
-          locale: Localizations.localeOf(context).toLanguageTag());
-    } catch (e) {
-      numFormat = NumberFormat.compact(locale: 'en');
-    }
-    final dataLength = numFormat.format(osmData.length);
-    final obsoleteDataLength = numFormat.format(osmData.obsoleteLength);
 
     return Scaffold(
       appBar: AppBar(title: Text(loc.settingsTitle)),
@@ -100,38 +83,12 @@ class SettingsPage extends ConsumerWidget {
                   },
                 ),
                 SettingsTile(
-                  title: Text(purgeAll
-                      ? loc.settingsPurgeAllData
-                      : loc.settingsPurgeData),
-                  trailing: Text(purgeAll ? dataLength : obsoleteDataLength),
-                  enabled: osmData.length > 0,
-                  onPressed: (_) async {
-                    if (purgeAll) {
-                      final answer = await showOkCancelAlertDialog(
-                        context: context,
-                        title: loc.settingsPurgeDataTitle,
-                        message: loc.settingsPurgeDataMessage,
-                        isDestructiveAction: true,
-                        okLabel: loc.buttonYes,
-                        cancelLabel: loc.buttonNo,
-                      );
-                      if (answer != OkCancelResult.ok) return;
-                    }
-
-                    int count =
-                        await ref.read(osmDataProvider).purgeData(purgeAll);
-                    if (purgeAll) {
-                      // Also delete unmodified notes.
-                      count += await ref
-                          .read(notesProvider)
-                          .purgeNotes(DateTime.now());
-                      AlertController.show(loc.settingsPurgedAllTitle,
-                          loc.settingsPurgedMessage(count), TypeAlert.success);
-                    } else {
-                      AlertController.show(loc.settingsPurgedObsoleteTitle,
-                          loc.settingsPurgedMessage(count), TypeAlert.success);
-                    }
-                    ref.read(needMapUpdateProvider).trigger();
+                  title: Text(loc.settingsDataManagement),
+                  onPressed: (context) async {
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(builder: (context) => CachesPage()),
+                    );
                   },
                 ),
               ],