+++ lib/screens/settings/imagery.dart
@@ -1,7 +1,6 @@
 import 'package:every_door/models/imagery.dart';
 import 'package:every_door/providers/imagery.dart';
 import 'package:every_door/providers/location.dart';
-import 'package:every_door/providers/tile_cacher.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:every_door/generated/l10n/app_localizations.dart' show AppLocalizations;
@@ -27,30 +26,11 @@ class ImageryPageState extends ConsumerState {
   @override
   Widget build(BuildContext context) {
     final imagery = ref.watch(imageryProvider);
-    final tileCacherState = ref.watch(tileCacheProvider);
     final loc = AppLocalizations.of(context)!;
 
     return Scaffold(
       appBar: AppBar(
         title: Text(loc.settingsBackground),
-        actions: [
-          IconButton(
-            onPressed: () {
-              final prov = ref.read(tileCacheProvider.notifier);
-              if (tileCacherState.idle)
-                prov.cacheForAll();
-              else
-                prov.stop();
-            },
-            icon: tileCacherState.idle
-                ? Icon(Icons.download)
-                : CircularProgressIndicator(
-                    value: tileCacherState.progress,
-                    color: Colors.white,
-                  ),
-            tooltip: loc.settingsCacheTiles,
-          ),
-        ],
       ),
       body: FutureBuilder(
         future: imageryList,